refactor(ios): MIM-68 split token activity view
